add test for empty index
